refactor: move dump to AbstractDiffMessage
