+++ core/src/main/java/org/bitcoinj/evolution/AbstractDiffMessage.java
@@ -19,6 +19,15 @@ package org.bitcoinj.evolution;
 
 import org.bitcoinj.core.Message;
 import org.bitcoinj.core.NetworkParameters;
+import org.bitcoinj.core.Utils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
 
 /**
  * Base class for all masternode/quorum list difference messages
@@ -27,10 +36,7 @@ import org.bitcoinj.core.NetworkParameters;
  */
 
 public abstract class AbstractDiffMessage extends Message {
-
-    public AbstractDiffMessage() {
-        super();
-    }
+    private static final Logger log = LoggerFactory.getLogger(AbstractDiffMessage.class);
 
     public AbstractDiffMessage(NetworkParameters params) {
         super(params);
@@ -38,5 +44,23 @@ public abstract class AbstractDiffMessage extends Message {
 
     public AbstractDiffMessage(NetworkParameters params, byte [] payload, int offset) {
         super(params, payload, offset);
+    }
+
+    protected abstract String getShortName();
+
+    public void dump(long startHeight, long endHeight) {
+        if (!Utils.isAndroidRuntime()) {
+            try {
+                File dumpFile = new File(getShortName() + "-" + params.getNetworkName() + "-" + startHeight + "-" + endHeight + ".dat");
+                OutputStream stream = new FileOutputStream(dumpFile);
+                stream.write(bitcoinSerialize());
+                stream.close();
+                log.info("dump successful");
+            } catch (FileNotFoundException x) {
+                log.warn("could not dump {} - file not found.", getShortName(), x);
+            } catch (IOException x) {
+                log.warn("could not dump {} - I/O error", getShortName(), x);
+            }
+        }
     }
 }
